Factor fft functions into new fft module
